create start/end nodes with campaign
