+++ packages/cm-models/src/lsst/cmservice/models/api/primitives.py
@@ -1,0 +1,8 @@
+"""Primitive models or skeletons for api models"""
+
+STEP_MANIFEST_TEMPLATE: dict[str, str | dict] = {
+    "apiVersion": "io.lsst.cmservice/v1",
+    "kind": "node",
+    "metadata": {},
+    "spec": {},
+}
